Move context store to its own file
